+++ __mocks__/gatsby.js
@@ -1,5 +1,4 @@
 const React = require("react");
-
 const gatsby = jest.requireActual("gatsby");
 
 module.exports = {
@@ -7,7 +6,6 @@ module.exports = {
   graphql: jest.fn(),
   Link: jest.fn().mockImplementation(
     // these props are invalid for an `a` tag
-    // eslint-disable-next-line no-unused-vars
     ({
       activeClassName,
       activeStyle,
