+++ gatsby-config.ts
@@ -119,7 +119,6 @@ const config: GatsbyConfig = {
     gatsbyPluginPageCreator,
     gatsbyPluginSitemap,
     "gatsby-plugin-sass",
-    "gatsby-plugin-react-helmet",
     "gatsby-plugin-image",
     "gatsby-plugin-sharp",
     "gatsby-transformer-sharp",
