+++ gatsby-ssr.js
@@ -5,3 +5,7 @@
  */
 
 // You can delete this file if you're not using it
+
+exports.onRenderBody = ({ setHtmlAttributes }) => {
+  setHtmlAttributes({ lang: "en" });
+};
