+++ package.json
@@ -12,7 +12,6 @@
     "gatsby-plugin-image": "^2.19.0",
     "gatsby-plugin-manifest": "^4.19.0",
     "gatsby-plugin-preact": "^6.19.0",
-    "gatsby-plugin-react-helmet": "^5.19.0",
     "gatsby-plugin-sass": "^5.19.0",
     "gatsby-plugin-sharp": "^4.19.0",
     "gatsby-plugin-sitemap": "^5.19.0",
@@ -29,7 +28,6 @@
     "preact-render-to-string": "5.2.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-helmet": "^6.1.0",
     "react-select": "^5.4.0",
     "remark": "^13.0.0",
     "sass": "^1.54.0",
