+++ src/components/HeadBuilder/index.ts
@@ -1,0 +1,5 @@
+import HeadBuilder from "./HeadBuilder";
+
+export type { QueryResult } from "./HeadBuilder";
+
+export default HeadBuilder;
