+++ src/components/OverratedDisappointmentsPage/OverratedDisappointmentsPage.spec.tsx
@@ -1,9 +1,10 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { act, render, screen, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import React from "react";
 import selectEvent from "react-select-event";
-import OverratedDisappointmentsPage from "./OverratedDisappointmentsPage";
+import OverratedDisappointmentsPage, {
+  Head,
+} from "./OverratedDisappointmentsPage";
 import data from "./OverratedDisappointmentsPage.fixtures";
 
 describe("/reviews/underseen/", () => {
@@ -13,17 +14,10 @@ describe("/reviews/underseen/", () => {
     expect(asFragment()).toMatchSnapshot();
   });
 
-  // Helmet uses requestAnimationFrame to ensure DOM is synced.
-  // https://github.com/nfl/react-helmet/blob/master/test/HelmetDeclarativeTest.js
-  // eslint-disable-next-line jest/no-done-callback
-  it("sets page title", (done) => {
-    expect.hasAssertions();
-    render(<OverratedDisappointmentsPage data={data} />);
+  it("sets page title", () => {
+    render(<Head />);
 
-    requestAnimationFrame(() => {
-      expect(document.title).toStrictEqual("Overrated Disappointments");
-      done();
-    });
+    expect(document.title).toStrictEqual("Overrated Disappointments");
   });
 
   it("can filter by title", async () => {
