--- src/components/Seo/index.ts
@@ -1,5 +1,0 @@
-import Seo from "./Seo";
-
-export type { SeoQueryResult } from "./Seo";
-
-export default Seo;
