+++ src/components/WatchlistEntityIndexPage/WatchlistEntityIndexPage.tsx
@@ -1,12 +1,12 @@
 import { graphql, Link } from "gatsby";
 import { GatsbyImage, IGatsbyImageData } from "gatsby-plugin-image";
-import React, { useReducer } from "react";
+import { useReducer } from "react";
 import DebouncedInput from "../DebouncedInput";
 import Fieldset from "../Fieldset";
 import FilterPageHeader from "../FilterPageHeader";
+import HeadBuilder from "../HeadBuilder";
 import Layout from "../Layout";
 import { SelectField } from "../SelectField";
-import Seo from "../Seo";
 import {
   containerCss,
   defaultImageCss,
@@ -159,6 +159,23 @@ function detailsForEntityType(entityType: EntityType) {
   }
 }
 
+export function Head({
+  pageContext,
+}: {
+  pageContext: PageContext;
+}): JSX.Element {
+  const entityDetails = detailsForEntityType(pageContext.entityType);
+
+  return (
+    <HeadBuilder
+      pageTitle={`Watchlist ${entityDetails.pluralName}`}
+      description={`A sortable and filterable list of watchlist ${entityDetails.pluralName.toLocaleLowerCase()}.`}
+      image={null}
+      article={false}
+    />
+  );
+}
+
 /**
  * Renders an index page for watchlist entities.
  */
@@ -181,12 +198,6 @@ export default function WatchlistEntityIndexPage({
 
   return (
     <Layout>
-      <Seo
-        pageTitle={`Watchlist ${entityDetails.pluralName}`}
-        description={`A sortable and filterable list of watchlist ${entityDetails.pluralName.toLocaleLowerCase()}.`}
-        image={null}
-        article={false}
-      />
       <main className={containerCss}>
         <div className={leftCss}>
           <FilterPageHeader
