+++ src/components/WatchlistEntityPage/WatchlistEntityPage.spec.tsx
@@ -1,8 +1,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { act, render, screen, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import React from "react";
-import WatchlistEntityPage, { EntityType } from "./WatchlistEntityPage";
+import WatchlistEntityPage, { EntityType, Head } from "./WatchlistEntityPage";
 import data, { dataNoAvatar } from "./WatchlistEntityPage.fixtures";
 
 describe("/watchlist/directors/{slug}", () => {
@@ -38,22 +37,12 @@ describe("/watchlist/directors/{slug}", () => {
     spy.mockRestore();
   });
 
-  // Helmet uses requestAnimationFrame to ensure DOM is synced.
-  // https://github.com/nfl/react-helmet/blob/master/test/HelmetDeclarativeTest.js
-  // eslint-disable-next-line jest/no-done-callback
-  it("sets page title", (done) => {
-    expect.hasAssertions();
+  it("sets page title", () => {
     render(
-      <WatchlistEntityPage
-        data={data}
-        pageContext={{ entityType: EntityType.DIRECTOR }}
-      />
+      <Head data={data} pageContext={{ entityType: EntityType.DIRECTOR }} />
     );
 
-    requestAnimationFrame(() => {
-      expect(document.title).toStrictEqual("John Ford");
-      done();
-    });
+    expect(document.title).toStrictEqual("John Ford");
   });
 
   it("can filter by title", async () => {
