+++ src/components/WatchlistEntityPage/WatchlistEntityPage.tsx
@@ -1,15 +1,15 @@
 import { graphql, Link } from "gatsby";
 import { IGatsbyImageData } from "gatsby-plugin-image";
-import React, { useReducer } from "react";
+import { useReducer } from "react";
 import Button from "../Button";
 import DebouncedInput from "../DebouncedInput";
 import Fieldset from "../Fieldset";
 import FilterPageHeader from "../FilterPageHeader";
+import HeadBuilder from "../HeadBuilder";
 import Layout from "../Layout";
 import { Poster, PosterList } from "../PosterList";
 import ProgressGraph from "../ProgressGraph";
 import { SelectField } from "../SelectField";
-import Seo from "../Seo";
 import YearInput from "../YearInput";
 import {
   containerCss,
@@ -162,6 +162,31 @@ function detailsForEntityType(
   }
 }
 
+export function Head({
+  pageContext,
+  data,
+}: {
+  pageContext: PageContext;
+  data: PageQueryResult;
+}): JSX.Element {
+  const entity = data.entity.nodes[0];
+
+  const entityDetails = detailsForEntityType(
+    pageContext.entityType,
+    entity.name,
+    entity.watchlistMovies.length
+  );
+
+  return (
+    <HeadBuilder
+      pageTitle={entity.name}
+      description={entityDetails.description}
+      image={null}
+      article={false}
+    />
+  );
+}
+
 /**
  * Renders a page for a watchlist director.
  */
@@ -199,12 +224,6 @@ export default function WatchlistEntityPage({
 
   return (
     <Layout>
-      <Seo
-        pageTitle={entity.name}
-        description={entityDetails.description}
-        image={null}
-        article={false}
-      />
       <main className={containerCss}>
         <div className={leftCss}>
           <FilterPageHeader
