+++ src/components/WatchlistIndexPage/WatchlistIndexPage.spec.tsx
@@ -1,8 +1,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { act, render, screen, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import React from "react";
-import WatchlistIndexPage from "./WatchlistIndexPage";
+import WatchlistIndexPage, { Head } from "./WatchlistIndexPage";
 import data from "./WatchlistIndexPage.fixtures";
 
 describe("/watchlist", () => {
@@ -12,17 +11,10 @@ describe("/watchlist", () => {
     expect(asFragment()).toMatchSnapshot();
   });
 
-  // Helmet uses requestAnimationFrame to ensure DOM is synced.
-  // https://github.com/nfl/react-helmet/blob/master/test/HelmetDeclarativeTest.js
-  // eslint-disable-next-line jest/no-done-callback
-  it("sets page title", (done) => {
-    expect.hasAssertions();
-    render(<WatchlistIndexPage data={data} />);
+  it("sets page title", () => {
+    render(<Head />);
 
-    requestAnimationFrame(() => {
-      expect(document.title).toStrictEqual("Watchlist");
-      done();
-    });
+    expect(document.title).toStrictEqual("Watchlist");
   });
 
   it("can filter by title", async () => {
