+++ src/pages/about.spec.tsx
@@ -1,7 +1,6 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { render } from "@testing-library/react";
-import React from "react";
-import AboutPage from "./about";
+import AboutPage, { Head } from "./about";
 import data from "./about.fixtures";
 
 describe("/about", () => {
@@ -11,16 +10,9 @@ describe("/about", () => {
     expect(asFragment()).toMatchSnapshot();
   });
 
-  // Helmet uses requestAnimationFrame to ensure DOM is synced.
-  // https://github.com/nfl/react-helmet/blob/master/test/HelmetDeclarativeTest.js
-  // eslint-disable-next-line jest/no-done-callback
-  it("sets page title", (done) => {
-    expect.hasAssertions();
-    render(<AboutPage data={data} />);
+  it("sets page title", () => {
+    render(<Head />);
 
-    requestAnimationFrame(() => {
-      expect(document.title).toStrictEqual("About");
-      done();
-    });
+    expect(document.title).toEqual("About");
   });
 });
