+++ src/pages/gone.tsx
@@ -1,27 +1,29 @@
 import { graphql } from "gatsby";
 import { IGatsbyImageData } from "gatsby-plugin-image";
-import React from "react";
 import ArticlePage from "../components/ArticlePage";
-import Seo from "../components/Seo";
+import HeadBuilder from "../components/HeadBuilder";
+
+export function Head(): JSX.Element {
+  return (
+    <HeadBuilder
+      pageTitle="410: Gone"
+      description="Forget it, Jake. It's Chinatown."
+      image={null}
+      article
+    />
+  );
+}
 
 export default function GonePage({ data }: PageQueryResult): JSX.Element {
   const { backdrop, page } = data;
 
   return (
-    <>
-      <Seo
-        pageTitle="410: Gone"
-        description="Forget it, Jake. It's Chinatown."
-        image={null}
-        article
-      />
-      <ArticlePage
-        image={backdrop.childImageSharp.gatsbyImageData}
-        alt="Jake Gittes walks away."
-        articleText={page.html}
-        title={page.frontmatter.title}
-      />
-    </>
+    <ArticlePage
+      image={backdrop.childImageSharp.gatsbyImageData}
+      alt="Jake Gittes walks away."
+      articleText={page.html}
+      title={page.frontmatter.title}
+    />
   );
 }
