+++ src/pages/how-i-grade.spec.tsx
@@ -1,7 +1,6 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { render } from "@testing-library/react";
-import React from "react";
-import HowIGradePage from "./how-i-grade";
+import HowIGradePage, { Head } from "./how-i-grade";
 import data from "./how-i-grade.fixtures";
 
 describe("/how-i-grade", () => {
@@ -11,16 +10,9 @@ describe("/how-i-grade", () => {
     expect(asFragment()).toMatchSnapshot();
   });
 
-  // Helmet uses requestAnimationFrame to ensure DOM is synced.
-  // https://github.com/nfl/react-helmet/blob/master/test/HelmetDeclarativeTest.js
-  // eslint-disable-next-line jest/no-done-callback
-  it("sets page title", (done) => {
-    expect.hasAssertions();
-    render(<HowIGradePage data={data} />);
+  it("sets page title", () => {
+    render(<Head />);
 
-    requestAnimationFrame(() => {
-      expect(document.title).toStrictEqual("How I Grade");
-      done();
-    });
+    expect(document.title).toStrictEqual("How I Grade");
   });
 });
